Move more tests to the core crate
